+++ CI/SCRIPTS/test_rasters.py
@@ -500,3 +500,27 @@ def test_dem_fct():
         ci.assert_raster_almost_equal(slope_path, slope_path_out, decimal=4)
         ci.assert_raster_almost_equal(slope_r_path, slope_r_path_out, decimal=4)
         ci.assert_raster_almost_equal(slope_p_path, slope_p_path_out, decimal=4)
+
+
+@s3_env
+def test_rasterize():
+    """ Test rasterize fct """
+    vec_path = rasters_path().joinpath("vector.geojson")
+    raster_path = rasters_path().joinpath("raster.tif")
+    raster_true_bin_path = rasters_path().joinpath("rasterized_bin.tif")
+    raster_true_path = rasters_path().joinpath("rasterized.tif")
+
+    with tempfile.TemporaryDirectory() as tmp_dir:
+        # Binary vector
+        out_bin_path = os.path.join(tmp_dir, "out_bin.tif")
+        rast_bin = rasters.rasterize(raster_path, vec_path)
+        rasters.write(rast_bin, out_bin_path, dtype=np.uint8, nodata=255)
+
+        ci.assert_raster_almost_equal(raster_true_bin_path, out_bin_path, decimal=4)
+
+        # Vector
+        out_path = os.path.join(tmp_dir, "out.tif")
+        rast = rasters.rasterize(raster_path, vec_path, "raster_val", dtype=np.uint8)
+        rasters.write(rast, out_path, dtype=np.uint8, nodata=255)
+
+        ci.assert_raster_almost_equal(raster_true_path, out_path, decimal=4)